% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_single_variable.R
\name{code_single_variable}
\alias{code_single_variable}
\title{Generate code for a single variable}
\usage{
code_single_variable(variable, distribution = "normal")
}
\arguments{
\item{variable}{Vector to generate code for}

\item{distribution}{Distribution of variable}
}
\description{
Generate code for a single variable
}
