% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupop.R
\name{dupop}
\alias{dupop}
\title{Generate code to create duplicate population}
\usage{
dupop(data, id_col, group_cols, n, ...)
}
\arguments{
\item{data}{A dataframe of observations to mimic}

\item{id_col}{Column name for individual observation id's}

\item{group_cols}{Columns to cluster by (ordered hierarchically)}

\item{...}{other arguments passed by name to each variable}
}
\description{
This
}
\examples{

df <- tibble::tibble(
  id = 1:100,
  group = rep(1:4, each = 25),
  gender = sample(c("m", "f"), 100, replace = TRUE),
  age = sample(18:45, 100, replace = TRUE, prob = dnorm(18:45, 24, 6)),
  height = rnorm(100, 160, 20)
)

df |>
  dupop(id_col = id, group_cols = group, n = 1000)
}
